39: split callbacks into different modules, dynamically register signal callbacks
